smol structure change
