+++ src/main/java/eu/greev/dcbot/ticketsystem/service/TranscriptData.java
@@ -1,0 +1,70 @@
+package eu.greev.dcbot.ticketsystem.service;
+
+import eu.greev.dcbot.ticketsystem.entities.Edit;
+import eu.greev.dcbot.ticketsystem.entities.Message;
+import eu.greev.dcbot.ticketsystem.entities.Ticket;
+import org.jdbi.v3.core.Jdbi;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class TranscriptData {
+    private final Jdbi jdbi;
+
+    protected TranscriptData(Jdbi jdbi) {
+        this.jdbi = jdbi;
+    }
+
+    public void loadTranscript(Ticket ticket) {
+        ticket.setTranscript(new Transcript(loadMessages(ticket.getId())));
+    }
+
+    private List<Message> loadMessages(int ticketId) {
+        List<Message> messages = new ArrayList<>();
+
+        jdbi.withHandle(handle -> handle.createQuery("SELECT * FROM messages WHERE ticketID = ?")
+                .bind(0, ticketId)
+                .map((r, columnNumber, ctx) -> {
+                    Message message = new Message(
+                            r.getLong("messageID"),
+                            r.getString("content"),
+                            r.getString("author"),
+                            r.getLong("timeCreated"));
+                    boolean isDeleted = r.getBoolean("isDeleted");
+                    boolean isEdited = r.getBoolean("isEdited");
+
+                    message.setDeleted(isDeleted);
+
+                    if (isEdited) {
+                        message.setEdits(loadEdits(message.getId()));
+                    }
+                    return null;
+                })
+                .findFirst());
+        return messages;
+    }
+
+    private List<Edit> loadEdits(long messageId) {
+        return jdbi.withHandle(handle -> handle.createQuery("SELECT content, timeEdited FROM edits WHERE messageID = ?")
+                .bind(0, messageId)
+                .mapTo(Edit.class)
+                .list());
+    }
+
+    public void deleteTranscript(Ticket ticket) {
+        List<Message> messages = ticket.getTranscript().getMessages();
+        List<Message> messagesWithEdits = messages.stream().filter(m -> !m.getEdits().isEmpty()).toList();
+
+        if (!messagesWithEdits.isEmpty()) {
+            for (Message message : messagesWithEdits) {
+                jdbi.withHandle(handle -> handle.createUpdate("DELETE FROM edits WHERE messageID=?")
+                        .bind(0, message.getId())
+                        .execute());
+            }
+        }
+
+        jdbi.withHandle(handle -> handle.createUpdate("DELETE FROM messages WHERE ticketID=?")
+                .bind(0, ticket.getId())
+                .execute());
+    }
+}
